% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_returns.R
\name{calc_returns}
\alias{calc_returns}
\title{Calculate compound investment growth}
\usage{
calc_returns(principal, rate, time, compound_freq = 1)
}
\arguments{
\item{principal}{Initial investment amount.}

\item{rate}{Annual interest rate as a decimal.}

\item{time}{Number of years.}

\item{compound_freq}{Number of compounding periods per year. Default is 1.}
}
\value{
A list containing the future value and a growth table.
}
\description{
Computes future value and a growth table for an investment.
}
